Separacion de modelos horario de dia y semana.
